Made some formatting changes to lib/log.lua and disabled buffering when trimming the log.
The reason for this is that log files can get really large, and
buffering loads the whole file into the buffer. We devised an algorithm
to not use too much memory when trimming log files by NOT loading
everything into RAM, but buffering nullified it. It has now been
disabled.